Added data entry

+++ bot.py
@@ -37,50 +37,47 @@ async def init(ctx):
 async def input(ctx):
     await ctx.message.delete()
     today = date.today()
-    # datajson = open(r"./data.json", 'r')
-    # jsonFile = json.loads(datajson.read())
-    # datajson.close()
-    # for element in jsonFile['elements']:
-    #     if element['user_id'] == ctx.author.id:
-    #         todayKey = str(today)
-    #         if(todayKey in element['dataRecord'].keys()):
-    #             todayData = element['dataRecord'][f'{today}']
     dateRecord = sf.checkRecord(ctx.author.id);
-                await ctx.send(f"Input already recorded today as '{todayData}'. Use *edit command to edit")
-                return
-    embed = discord.Embed(
-        title = f"Enter the values for {today}",
-        description ="||This request will timeout in 2 minuites||",
-        colour = discord.Colour.orange())
-    sent = await ctx.send(embed = embed)
-    def check(m):
-        if(ctx.author == m.author and m.channel == ctx.channel):
-            return True
-    try:
-        msg = await client.wait_for("message", timeout = 120, check = check)
-        if msg:
-            await sent.delete()
-            await msg.delete()
-            m = msg.content
-            def charCheck(s):
-                for i in s:
-                    if(i != 'y' and i != 'n'):
-                        return False
+    if dateRecord:
+        await ctx.send(f"Input already recorded today as '{dataRecord}'. Use *edit command to edit")
+        return
+    else:
+        embed = discord.Embed(
+            title = f"Enter the values for {today}",
+            description ="||This request will timeout in 2 minuites||",
+            colour = discord.Colour.orange())
+        sent = await ctx.send(embed = embed)
+        def check(m):
+            if(ctx.author == m.author and m.channel == ctx.channel):
                 return True
-            if(len(m)<=5 and charCheck(m)):
-                for i in range(len(jsonFile['elements'])):
-                    if(jsonFile['elements'][i]['user_id']==ctx.author.id):
-                        jsonFile['elements'][i]['dataRecord'][f'{today}'] = f'{m}'
-                        datajsonOut = open(r"./data.json", 'w')
-                        datajsonOut.write(json.dumps(jsonFile))
-                        datajsonOut.close()
-                        print(f"input updated : {jsonFile['elements'][i]}")
-                        await ctx.send(f"Input from {ctx.author.name} recorded as '{m}' for {today}. Very Naice")
-            else:
-                await ctx.send("Invalid input, type *input to try again")
-    except asyncio.TimeoutError:
-        await sent.delete()
-        await ctx.send("Cancelling due to timeout",delete_after=10)
+        try:
+            msg = await client.wait_for("message", timeout = 120, check = check)
+            if msg:
+                await sent.delete()
+                await msg.delete()
+                m = msg.content
+
+                def charCheck(s):
+                    for i in s:
+                        if(i != 'y' and i != 'n'):
+                            return False
+                    return True
+                if(len(m)<=5 and charCheck(m)):
+
+                    # for i in range(len(jsonFile['elements'])):
+                    #     if(jsonFile['elements'][i]['user_id']==ctx.author.id):
+                    #         jsonFile['elements'][i]['dataRecord'][f'{today}'] = f'{m}'
+                    #         datajsonOut = open(r"./data.json", 'w')
+                    #         datajsonOut.write(json.dumps(jsonFile))
+                    #         datajsonOut.close()
+                    sf.inputData(ctx.author.id, m)
+                    print(f"input updated : {m}")
+                    await ctx.send(f"Input from {ctx.author.name} recorded as '{m}' for {today}. Very Naice")
+                else:
+                    await ctx.send("Invalid input, type *input to try again")
+        except asyncio.TimeoutError:
+            await sent.delete()
+            await ctx.send("Cancelling due to timeout",delete_after=10)
 
 @client.command()
 async def edit(ctx):
